Remove attempt to migrate from legacy encoding.
